Simplify temperature conversion and update mapping comment
Agent-Logs-Url: https://github.com/MKaluz/RegisterApp/sessions/f013616d-5581-42c5-908e-532ec034f0a9

+++ RegisterApp.Api/Program.cs
@@ -40,5 +40,5 @@ app.Run();
 
 record WeatherForecast(DateOnly Date, int TemperatureC, string? Summary)
 {
-    public int TemperatureF => 32 + (int)(TemperatureC * 9 / 5.0);
+    public int TemperatureF => 32 + TemperatureC * 9 / 5;
 }
